feat: support relative path1 in join.
This is to support aurelia.use.feature('../src') and aurelia.use.plugin('../../local-plugin/dist/native-modules') for easier local plugin development. To be fair, this doesn't work in cli-bundler as requirejs doesn't support moduleId starting in "../". This doesn't work in webpack either, but we probably can update aurelia-webpack-plugin PreserveModuleNamePlugin to support it. This feature improves dumber experience when developing Aurelia plugin locally. Well, we don't have official offering of skeleton for dumber, but I am planning to use this feature in https://github.com/dumberjs/new.

closes #27